moving go.mod to root
