+++ go.mod
@@ -7,4 +7,11 @@ require (
 	github.com/labstack/echo/v4 v4.1.17
 	google.golang.org/grpc v1.33.2
 	google.golang.org/protobuf v1.25.0
+
+	github.com/go-logr/logr v0.1.0
+	github.com/onsi/ginkgo v1.11.0
+	github.com/onsi/gomega v1.8.1
+	k8s.io/apimachinery v0.17.2
+	k8s.io/client-go v0.17.2
+	sigs.k8s.io/controller-runtime v0.5.0
 )
