--- operator-old/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "operator"
-version = "0.1.0"
-authors = ["Aaron Schlesinger <aaron@ecomaz.net>"]
-edition = "2018"
-
-[dependencies]
-kube = "0.42.0"
-kube-derive = "0.43.0"
-kube-runtime = "0.42.0"
-k8s-openapi = { version = "0.9.0", default-features = false, features = ["v1_17"] }
-serde = "1.0"
-serde_derive = "1.0"
-serde_json = "1.0"
-tokio = { version = "0.2.22", features = ["macros"] }
-snafu = { version = "0.6.8", features = ["futures"] }
-futures = "0.3.5"
-log = "0.4.11"
-
-
