--- operator-old/Dockerfile
@@ -1,35 +1,0 @@
-# see background on cargo-chef:
-# https://www.lpalmieri.com/posts/fast-rust-docker-builds/
-
-
-FROM rust as planner
-
-WORKDIR /app
-RUN cargo install cargo-chef 
-COPY . .
-RUN cargo chef prepare  --recipe-path recipe.json
-
-
-FROM rust as cacher
-
-WORKDIR /app
-RUN cargo install cargo-chef
-COPY --from=planner /app/recipe.json recipe.json
-RUN cargo chef cook --release --recipe-path recipe.json
-
-
-FROM rust as builder
-
-WORKDIR /app
-COPY . .
-# Copy over the cached dependencies
-COPY --from=cacher /app/target target
-COPY --from=cacher /usr/local/cargo /usr/local/cargo
-RUN cargo build --release --bin operator
-
-
-FROM rust as runtime
-
-WORKDIR /app
-COPY --from=builder /app/target/release/app /usr/local/bin
-ENTRYPOINT ["./usr/local/bin/app"]
