--- operator-old/README.md
@@ -1,3 +1,0 @@
-# KEDA-HTTP Operator
-
-This operator is responsible for receiving `App` Custom Resource Definitions (CRDs) and creating and maintaining the necessary Kubernetes resources to run the application in the cluster.
